feat: implement payment order (orden_pago) mapping, including entity support, batch grouping, and provider lookup utilities.
